app: allow a dark variant of a same theme, instead of having 2 themes.
This way, what will happen is that:

- We can have a single "Default" theme which will have both the light
  and dark versions.
- With our Default theme, when "Use dark theme variant if available" is
  unchecked, we just follow the system-wide dark settings. (though I'm
  unsure we actually do with current code; we do load our theme over the
  system theme, which may be dark, but I don't think we'd load a dark
  theme variant then)
- If the option is checked, we will load the specific dark variant,
  bypassing system settings specifically for GIMP.

Technically for theme designers, all it takes to have a dark variant is
to add a gimp-dark.css next to gimp.css. `gimp-dark.css` is loaded
instead of `gimp.css` when the settings is checked.

Note: there is apparently a new freedesktop portal for setting the
prefered variant (and now it's apparently either light, dark or
default), which is now implemented by GNOME, KDE and Elementary at
least. It would be nice if we could grab this settings and use it if
available. The below link has code sample showing how to do it with
DBus:
https://gitlab.gnome.org/GNOME/Initiatives/-/wikis/Dark-Style-Preference