app: use our custom layer search entry with Ctrl-F in the Layers dockable.
GtkTreeView has an "interactive search" feature built-in, which basically kicks
in with the Ctrl-F shortcut by default. Let's hijack this feature and trigger
our own new search popover which is much more powerful (multi-item selection,
ability to use regexp or glob search by enabling these in Preferences, etc.).

This was an idea by Aryeom who wanted to be able to hit Ctrl-F to search for
layers.